cart and wishlist backend working

+++ This_App_Backend/src/main/java/com/example/This_App_Backend/entity/Wishlist.java
@@ -22,7 +22,7 @@ public class Wishlist {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "wishlist_id")
-    private Integer wishlistId;
+    private Long  wishlistId;
 
     // Many Wishlists items belong to one User
     @ManyToOne(fetch = FetchType.LAZY)
